// sdk/batch/batch/src/api/models.ts
export interface VMExtension {
  /** The name of the virtual machine extension. */
  name: string;
  /** The name of the extension handler publisher. */
  publisher: string;
  /** The type of the extension. */
  type: string;
  /** The version of script handler. */
  typeHandlerVersion?: string;
  /**
   * Indicates whether the extension should use a newer minor version if one is
   * available at deployment time. Once deployed, however, the extension will not
   * upgrade minor versions unless redeployed, even with this property set to true.
   */
  autoUpgradeMinorVersion?: boolean;
  /** JSON formatted public settings for the extension. */
  settings?: object;
  /**
   * The extension can contain either protectedSettings or
   * protectedSettingsFromKeyVault or no protected settings at all.
   */
  protectedSettings?: object;
  /**
   * Collection of extension names after which this extension needs to be
   * provisioned.
   */
  provisionAfterExtensions?: string[];
}
// ...
export interface NodeVMExtension {
  /** The provisioning state of the virtual machine extension. */
  provisioningState?: string;
  /** The configuration for virtual machine extensions. */
  vmExtension?: VMExtension;
  /** The vm extension instance view. */
  instanceView?: VMExtensionInstanceView;
}

/** The vm extension instance view. */
export interface VMExtensionInstanceView {
  /** The name of the vm extension instance view. */
  name?: string;
  /** The resource status information. */
  statuses?: InstanceViewStatus[];
  /** The resource status information. */
  subStatuses?: InstanceViewStatus[];
}

/** The instance view status. */
export interface InstanceViewStatus {
  /** The status code. */
  code?: string;
  /** The localized label for the status. */
  displayStatus?: string;
  /** Level code. */
  level?: StatusLevelTypes;
  /** The detailed status message. */
  message?: string;
  /** The time of the status. */
  time?: string;
}

/** */
/** "Error", "Info", "Warning" */
export type StatusLevelTypes = string;

/** */
export interface NodeVMExtensionList {
  /** The list of Compute Node extensions. */
  value?: NodeVMExtension[];
  /** The URL to get the next set of results. */
  nextLink?: string;
}